+++ tig-challenges/src/vector_search.rs
@@ -51,7 +51,7 @@ impl TryFrom<Map<String, Value>> for Solution {
 
 #[derive(Serialize, Deserialize, Debug)]
 pub struct Challenge {
-    pub seed: u32,
+    pub seed: u64,
     pub difficulty: Difficulty,
     pub vector_database: Vec<Vec<f32>>,
     pub query_vectors: Vec<Vec<f32>>,
@@ -73,7 +73,7 @@ pub const KERNEL: Option<CudaKernel> = None;
 impl ChallengeTrait<Solution, Difficulty, 2> for Challenge {
     #[cfg(feature = "cuda")]
     fn cuda_generate_instance(
-        seed: u32,
+        seed: u64,
         difficulty: &Difficulty,
         dev: &Arc<CudaDevice>,
         mut funcs: HashMap<&'static str, CudaFunction>,
@@ -82,8 +82,8 @@ impl ChallengeTrait<Solution, Difficulty, 2> for Challenge {
         Self::generate_instance(seed, difficulty)
     }
 
-    fn generate_instance(seed: u32, difficulty: &Difficulty) -> Result<Self> {
-        let mut rng = StdRng::seed_from_u64(seed as u64);
+    fn generate_instance(seed: u64, difficulty: &Difficulty) -> Result<Self> {
+        let mut rng = StdRng::seed_from_u64(seed);
         let uniform = Uniform::from(0.0..1.0);
         let search_vectors = (0..100000)
             .map(|_| (0..250).map(|_| uniform.sample(&mut rng)).collect())
